+++ Data/src/main/java/com/hch/mapper/ScrapApplyMapper.java
@@ -3,13 +3,13 @@ package com.hch.mapper;
 import com.hch.model.entity.ScrapApply;
 
 public interface ScrapApplyMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(ScrapApply record);
 
     int insertSelective(ScrapApply record);
 
-    ScrapApply selectByPrimaryKey(Long id);
+    ScrapApply selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(ScrapApply record);
