+++ Data/src/main/java/com/hch/mapper/ScrapMapper.java
@@ -3,13 +3,13 @@ package com.hch.mapper;
 import com.hch.model.entity.Scrap;
 
 public interface ScrapMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Scrap record);
 
     int insertSelective(Scrap record);
 
-    Scrap selectByPrimaryKey(Long id);
+    Scrap selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(Scrap record);
