+++ Data/src/main/java/com/hch/mapper/UseApplyMapper.java
@@ -3,13 +3,13 @@ package com.hch.mapper;
 import com.hch.model.entity.UseApply;
 
 public interface UseApplyMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(UseApply record);
 
     int insertSelective(UseApply record);
 
-    UseApply selectByPrimaryKey(Long id);
+    UseApply selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(UseApply record);
