+++ Data/src/main/java/com/hch/mapper/UseMapper.java
@@ -3,13 +3,13 @@ package com.hch.mapper;
 import com.hch.model.entity.Use;
 
 public interface UseMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Use record);
 
     int insertSelective(Use record);
 
-    Use selectByPrimaryKey(Long id);
+    Use selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(Use record);
